diverse fikser for aa faa M5 til aa kompilere paa stallo

+++ m5/mem/cache/miss/throttle_control.cc
@@ -151,7 +151,7 @@ ThrottleControl::useAveragePolicy(Tick time, int cpuid){
 Tick
 ThrottleControl::useTokenPolicy(Tick time, int cpuid){
 
-	int cyclesBetweenRequests = (1/targetRequestRate[cpuid]);
+        int cyclesBetweenRequests = (int) (1.0/targetRequestRate[cpuid]);
 	if(tokenRunLast[cpuid] < time){
 
 		Tick cyclesSinceLast = time - tokenRunLast[cpuid];
